FishMap-13 - First refactor - multiple sagas
