+++ src/actions/fishmarks.js
@@ -1,5 +1,5 @@
-import {SET_FISHMARK_POSITION, UPDATE_FISHMARK_DATA, MOVE_TO_FISHMARK_POSITION,
-    DELETE_FISHMARK_POSITION, LOAD_POSITIONS, UPLOAD_FISHMARK_POSITIONS, LOAD_WAYPOINTS_ON_PUSH,
+import {SET_FISHMARK_POSITION, MOVE_TO_FISHMARK_POSITION,LOAD_FISHMARKS_POSITIONS,
+    DELETE_FISHMARK_POSITION, LOAD_WAYPOINTS_ON_PUSH,
 } from '../constants/constants'
 
 export const setFishmark = (data) => ({
@@ -19,13 +19,8 @@ export const deleteFishmarkPosition = (position) => ({
 });
 
 export const loadFishPositions = () => ({
-    type: LOAD_POSITIONS
+    type: LOAD_FISHMARKS_POSITIONS
 });
-
-export const uploadFishmarkPositions = data => ({
-    type: UPLOAD_FISHMARK_POSITIONS,
-    data,
-})
 
 export const loadFishWaypointsOnPush = () =>({
     type: LOAD_WAYPOINTS_ON_PUSH
