+++ src/components/UserMarker.js
@@ -8,7 +8,6 @@ class UserMarker extends Component {
 
     render() {
 
-        console.log("Marker user", this.props.marker)
         const region = {
             latitude:this.props.marker.latitude,
             longitude: this.props.marker.longitude,
