+++ src/constants/constants.js
@@ -4,13 +4,13 @@ export const GET_FISHMARK_POSITION = "GET_FISHMARK_POSITION"
 export const UPDATE_FISHMARK_DATA = "UPDATE_FISHMARK_DATA"
 export const MOVE_TO_FISHMARK_POSITION = "MOVE_TO_FISHMARK_POSITION"
 export const DELETE_FISHMARK_POSITION = "DELETE_FISHMARK_POSITION"
-export const LOAD_POSITIONS = "LOAD_POSITIONS"
+export const LOAD_FISHMARKS_POSITIONS_PENDING = "LOAD_FISHMARKS_POSITIONS_PENDING"
 
-export const UPLOAD_FISHMARK_POSITIONS = "UPLOAD_FISHMARK_POSITIONS"
-export const FAILED_UPLOAD_FISHMARK_POSITIONS = "FAILED_UPLOAD_FISHMARK_POSITIONS"
+export const SUCCESS_LOAD_FISHMARKS_POSITIONS = "SUCCESS_LOAD_FISHMARKS_POSITIONS"
+export const FAILED_LOAD_FISHMARKS_POSITIONS = "FAILED_LOAD_FISHMARKS_POSITIONS"
 export const FAILED_SET_FISHMARK_POSITION = "FAILED_SET_FISHMARK_POSITION"
 export const SUCCESS_SET_FISHMARK_POSITION = "SUCCESS_SET_FISHMARK_POSITION"
-export const SUCCESS_UPLOAD_FISHMARK_POSITIONS = "SUCCESS_UPLOAD_FISHMARK_POSITIONS"
+export const LOAD_FISHMARKS_POSITIONS = "LOAD_FISHMARKS_POSITIONS"
 
 
 
@@ -26,6 +26,10 @@ export const LOGIN = "LOGIN"
 export const FAILED_SET_TOKEN = "FAILED_SET_TOKEN"
 export const SUCCESS_SET_TOKEN = "SUCCESS_SET_TOKEN"
 export const VERIFY_TOKEN = "VERIFY_TOKEN"
+export const SUCCESS_REMOVE_TOKEN = "SUCCESS_REMOVE_TOKEN"
+export const FAILED_REMOVE_TOKEN = "FAILED_REMOVE_TOKEN"
+export const SUCCESS_VERIFY_TOKEN = "SUCCESS_VERIFY_TOKEN"
+export const FAILED_VERIFY_TOKEN = "FAILED_VERIFY_TOKEN"
 
 export const SUCCESS_GET_USER_LOCATION = "SUCCESS_GET_USER_LOCATION"
 export const FAILED_GET_USER_LOCATION  = "FAILED_GET_USER_LOCATION"
@@ -33,7 +37,7 @@ export const GET_USER_LOCATION = "GET_USER_LOCATION"
 export const LOAD_WAYPOINTS_ON_PUSH = "LOAD_WAYPOINTS_ON_PUSH"
 export const SUCCESS_UPDATE_WAYPOINTS_ON_PUSH =  "SUCCESS_UPDATE_WAYPOINTS_ON_PUSH"
 export const FAILED_UPDATE_WAYPOINT_ON_PUSH = "FAILED_UPDATE_WAYPOINT_ON_PUSH"
-export const FINISHED_UPDATE_WAYPOINTS_ON_PUSH = "FINISHED_UPDATE_WAYPOINTS_ON_PUSH"
+export const INCREMENT_SEED = "INCREMENT_SEED"
 
 export const DELETE_WAYPOINT = "DELETE_WAYPOINT"
 export const SUCCESS_DELETE_WAYPOINT = "SUCCESS_DELETE_WAYPOINT"
