+++ src/reducers/user.js
@@ -1,11 +1,15 @@
-import {SET_USER_DATA, SUCCESS_SET_TOKEN, FAILED_SET_TOKEN, FAILED_GET_USER_LOCATION, SUCCESS_GET_USER_LOCATION} from '../constants/constants'
+import {
+    SET_USER_DATA, SUCCESS_SET_TOKEN, FAILED_SET_TOKEN, FAILED_GET_USER_LOCATION, SUCCESS_GET_USER_LOCATION,
+    VERIFY_TOKEN, SUCCESS_VERIFY_TOKEN
+} from '../constants/constants'
 import {setFishmark} from "../actions/fishmarks";
 
 const intialState = {
     user: {},
     success:false,
     error:false,
-    position:{}
+    position:{},
+    message:null
 }
 
 const reducer = (state=intialState, action) => {
@@ -31,6 +35,10 @@ const reducer = (state=intialState, action) => {
         case FAILED_GET_USER_LOCATION:
             return {
                 ...state, error:action.error
+            }
+        case SUCCESS_VERIFY_TOKEN:
+            return {
+                ...state
             }
         default:
             return state
